Move pytest config to pyproject
