+++ AGENTS.md
@@ -1,0 +1,4 @@
+# Repository Guidelines
+
+- Run `uv run ruff check --fix src tests` before committing to ensure code style.
+- Run `uv run pytest -q` to execute the test suite.
